// src/generated/models/Afb_utilisateurinternesModel.ts
/*!
 * Copyright (C) Microsoft Corporation. All rights reserved.
 * This file is autogenerated. Do not edit this file directly.
 */
export const Afb_utilisateurinternesafb_actif = {
  0: 'Oui',
  747010001: 'Non'
} as const;
export type Afb_utilisateurinternesafb_actif = keyof typeof Afb_utilisateurinternesafb_actif;
export const Afb_utilisateurinternesafb_direction = {
  747010000: 'DCONF',
  747010001: 'DMG',
  747010002: 'DSI',
  747010003: 'TRESO',
  747010004: 'DRISQUE'
} as const;
export type Afb_utilisateurinternesafb_direction = keyof typeof Afb_utilisateurinternesafb_direction;
export const Afb_utilisateurinternesafb_role = {
  747010000: 'Auditeurinterne',
  747010001: 'Charg_deconformit_',
  747010002: 'Charg_derelation',
  747010003: 'SuperadminDCONF',
  747010004: 'Auditeurexterne'
} as const;
export type Afb_utilisateurinternesafb_role = keyof typeof Afb_utilisateurinternesafb_role;
export const Afb_utilisateurinternesstatecode = {
  0: 'Actif',
  1: 'Inactif'
} as const;
export type Afb_utilisateurinternesstatecode = keyof typeof Afb_utilisateurinternesstatecode;
export const Afb_utilisateurinternesstatuscode = {
  1: 'Actif',
  2: 'Inactif'
} as const;
export type Afb_utilisateurinternesstatuscode = keyof typeof Afb_utilisateurinternesstatuscode;

export interface Afb_utilisateurinternesBase {
  afb_actif: Afb_utilisateurinternesafb_actif;
  afb_adresseemail: string;
  afb_derniereconnexion?: string;
  afb_direction: Afb_utilisateurinternesafb_direction;
  afb_droitsrbac?: string;
  afb_identifiantcisco?: string;
  afb_nomcomplet: string;
  afb_role: Afb_utilisateurinternesafb_role;
  afb_utilisateurinterneid: string;
  importsequencenumber?: number;
  overriddencreatedon?: string;
  ownerid: string;
  owneridtype: string;
  statecode: Afb_utilisateurinternesstatecode;
  statuscode?: Afb_utilisateurinternesstatuscode;
  timezoneruleversionnumber?: number;
  utcconversiontimezonecode?: number;
}

export interface Afb_utilisateurinternes extends Afb_utilisateurinternesBase {
  afb_actifname?: string;
  afb_directionname?: string;
  afb_rolename?: string;
  createdbyname?: string;
  createdbyyominame: string;
  createdon?: string;
  createdonbehalfbyname?: string;
  createdonbehalfbyyominame: string;
  modifiedbyname?: string;
  modifiedbyyominame: string;
  modifiedon?: string;
  modifiedonbehalfbyname?: string;
  modifiedonbehalfbyyominame: string;
  owneridname: string;
  owneridyominame: string;
  owningbusinessunitname: string;
  statecodename?: string;
  statuscodename?: string;
  versionnumber?: number;
  createdby?: object;
  _createdby_value?: string;
  createdonbehalfby?: object;
  _createdonbehalfby_value?: string;
  modifiedby?: object;
  _modifiedby_value?: string;
  modifiedonbehalfby?: object;
  _modifiedonbehalfby_value?: string;
  owningbusinessunit?: object;
  _owningbusinessunit_value?: string;
  owningteam?: object;
  _owningteam_value?: string;
  owninguser?: object;
  _owninguser_value?: string;
}